test: reorganized folders in test.service
