+++ src/test/java/ulaval/glo2003/service/unit/SellingServiceTest.java
@@ -1,4 +1,4 @@
-package ulaval.glo2003.service;
+package ulaval.glo2003.service.unit;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -24,6 +24,10 @@ import ulaval.glo2003.domain.product.ProductTestUtils;
 import ulaval.glo2003.domain.seller.ISellerRepository;
 import ulaval.glo2003.domain.seller.Seller;
 import ulaval.glo2003.domain.seller.SellerTestUtils;
+import ulaval.glo2003.service.OfferMapper;
+import ulaval.glo2003.service.ProductMapper;
+import ulaval.glo2003.service.SellerMapper;
+import ulaval.glo2003.service.SellingService;
 
 class SellingServiceTest {
     private static final String SELLER_ID = "SELLER";
